fix typescript compatibility
